requirements.txt
Flask
pyradios
python-mpv
requests
